Completed bash function for backing up my MSc files to another partition and to a ssh server on the university

+++ bashrc
@@ -121,45 +121,65 @@ eval `dircolors ~/.dir_colors/dircolors-solarized/dircolors.ansi-dark`
 
 MSc="$HOME/Dokument/University_of_Sussex/MScProject"
 
-function backup()
-{### NEED TO MAKE IT WORK WHEN HARD DRIVE IS MOUNTED
-Storage=/media/kristian/storage
+function BackupMSc()
+{
+    mount_point=/dev/sda5
+    
+    if mount | grep "$mount_point" > /dev/null;
+    then
+        Storage=$(mount | grep $mount_point | grep -Po '(on\s)\K[^\s]*')
+        unmount=false
+    else
+        Storage=/media/kristian/storage
+        unmount=true
+    
+        sudo mkdir $Storage
+        sudo mount /dev/sda5 $Storage
+    fi
+    
+    DirToBackup=/home/kristian/Dokument/University_of_Sussex/MScProject/
+    BackupDir="$Storage/Dokumenter/MScBackup"
+    
+    
+    timestamp=$(date "+%Y-%m-%d_%H-%M-%S")
+    
+    sudo install --directory "$BackupDir/history/$timestamp"
+    sudo install --directory "$BackupDir/backup/"
+    sudo install --directory "$BackupDir/log/$timestamp"
+    
+    echo "Backup $timestamp"
+    
+    rsync --dry-run --itemize-changes --out-format="%i|%n|"  --recursive --update --delete --perms --owner --group --times --links --safe-links --super --one-file-system --devices $DirToBackup "$BackupDir/backup" | sed '/^ *$/d' > "$BackupDir/log/$timestamp/dryrun.log"
+    
+    grep "^.f" "$BackupDir/log/$timestamp/dryrun.log" >> "$BackupDir/log/$timestamp/onlyfiles.log"
+    grep "^.f+++++++++" "$BackupDir/log/$timestamp/onlyfiles.log" | awk -F '|' '{print $2 }' | sed 's@^/@@' >> "$BackupDir/log/$timestamp/created.log"
+    grep --invert-match "^.f+++++++++" "$BackupDir/log/$timestamp/onlyfiles.log" | grep --invert-match "^.f\.\.\.pog\.\.\." | awk -F '|' '{print $2 }' | sed 's@^/@@' >> "$BackupDir/log/$timestamp/changed.log"
+    #grep --invert-match "^.f+++++++++" "$BackupDir/log/$timestamp/onlyfiles.log" | awk -F '|' '{print $2 }' | sed 's@^/@@' >> "$BackupDir/log/$timestamp/changed.log"
+    
+    grep "^\.d" "$BackupDir/log/$timestamp/dryrun.log" | grep --invert-match "^.d\.\.\.pog\.\.\." | awk -F '|' '{print $2 }' | sed -e 's@^/@@' -e 's@^\./@@' -e 's@/$@@' >> "$BackupDir/log/$timestamp/changed.log"
+    grep "^cd" "$BackupDir/log/$timestamp/dryrun.log" | awk -F '|' '{print $2 }' | sed -e 's@^/@@' -e 's@/$@@' >> "$BackupDir/log/$timestamp/created.log"
+    
+    grep "^*deleting" "$BackupDir/log/$timestamp/dryrun.log" | awk -F '|' '{print $2 }' >> "$BackupDir/log/$timestamp/deleted.log"
+    
+    cat "$BackupDir/log/$timestamp/deleted.log" > /tmp/tmp.rsync.list
+    cat "$BackupDir/log/$timestamp/changed.log" >> /tmp/tmp.rsync.list
+    sort --output=/tmp/rsync.list --unique /tmp/tmp.rsync.list
+    
+    rsync --update --perms --owner --group --times --links --super --files-from=/tmp/rsync.list "$BackupDir/backup" "$BackupDir/history/$timestamp/"
+    
+    rsync --recursive --update --delete --perms --owner --group --times --links --safe-links --super --one-file-system --devices $DirToBackup "$BackupDir/backup"
+    
+    if mount | grep "$mount_point" > /dev/null;
+    then
+        if $unmount; then
+            sudo umount $Storage
+            sudo rm -r $Storage
+        fi
+    fi
 
-sudo mkdir $Storage
-sudo mount /dev/sda5 $Storage
+    echo "Main backup at: $mount_point/Dokumenter/MScBackup"
+    
+    rsync --archive --update --verbose --compress --rsh=ssh $DirToBackup feynman:MScBackup
 
-DirToBackup=/home/kristian/test/
-BackupDir="$Storage/Dokumenter/MScBackup"
-
-
-timestamp=$(date "+%Y-%m-%d_%H-%M-%S")
-
-sudo install --directory "$BackupDir/history/$timestamp"
-sudo install --directory "$BackupDir/backup/"
-sudo install --directory "$BackupDir/log/$timestamp"
-
-rsync --dry-run --itemize-changes --out-format="%i|%n|"  --recursive --update --delete --perms --owner --group --times --links --safe-links --super --one-file-system --devices $DirToBackup "$BackupDir/backup" | sed '/^ *$/d' > "$BackupDir/log/$timestamp/dryrun.log"
-
-grep "^.f" "$BackupDir/log/$timestamp/dryrun.log" >> "$BackupDir/log/$timestamp/onlyfiles.log"
-grep "^.f+++++++++" "$BackupDir/log/$timestamp/onlyfiles.log" | awk -F '|' '{print $2 }' | sed 's@^/@@' >> "$BackupDir/log/$timestamp/created.log"
-grep --invert-match "^.f+++++++++" "$BackupDir/log/$timestamp/onlyfiles.log" | grep --invert-match "^.f\.\.\.pog\.\.\." | awk -F '|' '{print $2 }' | sed 's@^/@@' >> "$BackupDir/log/$timestamp/changed.log"
-#grep --invert-match "^.f+++++++++" "$BackupDir/log/$timestamp/onlyfiles.log" | awk -F '|' '{print $2 }' | sed 's@^/@@' >> "$BackupDir/log/$timestamp/changed.log"
-
-grep "^\.d" "$BackupDir/log/$timestamp/dryrun.log" | grep --invert-match "^.d\.\.\.pog\.\.\." | awk -F '|' '{print $2 }' | sed -e 's@^/@@' -e 's@^\./@@' -e 's@/$@@' >> "$BackupDir/log/$timestamp/changed.log"
-grep "^cd" "$BackupDir/log/$timestamp/dryrun.log" | awk -F '|' '{print $2 }' | sed -e 's@^/@@' -e 's@/$@@' >> "$BackupDir/log/$timestamp/created.log"
-
-grep "^*deleting" "$BackupDir/log/$timestamp/dryrun.log" | awk -F '|' '{print $2 }' >> "$BackupDir/log/$timestamp/deleted.log"
-
-cat "$BackupDir/log/$timestamp/deleted.log" > /tmp/tmp.rsync.list
-cat "$BackupDir/log/$timestamp/changed.log" >> /tmp/tmp.rsync.list
-sort --output=/tmp/rsync.list --unique /tmp/tmp.rsync.list
-
-rsync --update --perms --owner --group --times --links --super --files-from=/tmp/rsync.list "$BackupDir/backup" "$BackupDir/history/$timestamp/"
-
-rsync --recursive --update --delete --perms --owner --group --times --links --safe-links --super --one-file-system --devices $DirToBackup "$BackupDir/backup"
-
-sudo umount $Storage
-sudo rm -r $Storage
-
-rsync --archive --update --verbose --compress --rsh=ssh $DirToBackup feynman:backup
+    echo "Secondary backup at: kb339@feynman.hpc.susx.ac.uk:MScBackup"
 }
